Peripheral: Added EXTI handler
Added EXTI4 handler and global interrupts enabled

Closes #1

+++ system_functions.c
@@ -4,6 +4,6 @@
 
 void enableIRQ(int irqNumber)
 {
-  NVIC->ISER[(((uint32_t)(int32_t)irqNumber) >> 5UL)] = (uint32_t)(1UL << (((uint32_t)(int32_t)irqNumber) & 0x1FUL));
+  NVIC->ISER[(((uint32_t)irqNumber) >> 5UL)] = (uint32_t)(1UL << (((uint32_t)(int32_t)irqNumber) & 0x1FUL));
 }
